Added sales order branch

+++ API/Dtos/ProductGemDto.cs
@@ -5,19 +5,19 @@
         public int GemTypeId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Proportion { get; set; }
+        public decimal Proportion { get; set; }
         public string Polish { get; set; }
         public string Fluorescence { get; set; }
         public string Symmetry { get; set; }
-        public float Carat { get; set; }
+        public decimal Carat { get; set; }
         public string Cut { get; set; }
         public string Clarity { get; set; }
         public string Color { get; set; }
         public string Shape { get; set; }
-        public float LatestPrice { get; set; }
-        public float GemWeight { get; set; }
+        public decimal LatestPrice { get; set; }
+        public decimal GemWeight { get; set; }
         public string CertificateCode { get; set; }
         public int Quantity { get; set; }
-        public float GemsPrice { get; set; }
+        public decimal GemsPrice { get; set; }
     }
 }
