+++ Core/Enitities/GoldType.cs
@@ -8,9 +8,9 @@ namespace Core.Enitities
         [Column(TypeName = "nvarchar(100)"), Required]
         public string Name { get; set; }
         [Column(TypeName = "decimal(18,2)"), Required]
-        public float LatestBidPrice { get; set; }
+        public decimal LatestBidPrice { get; set; }
         [Column(TypeName = "decimal(18,2)"), Required]
-        public float LatestAskPrice { get; set; }
+        public decimal LatestAskPrice { get; set; }
         [Column(TypeName = "nvarchar(10)"), Required]
         public string Unit { get; set; }
         public bool Status { get; set; } = true;
